Polish diagnostics detail pane architecture
